IMPALA-11970: Add query timeline display to Impala WebUI
This change adds a chart to the Impala WebUI for displaying query
fragments and plan node timings. The chart list execution timings in a
compact form with the plan/fragment tree shown on the left side and a
Gantt chart with bars for each execution phase on the right side.

Plan node labels are color-coded by fragment. The node list will scroll
for large plans so that the scale and legend always stay visible.
Exchange nodes have dotted boxes that show when rows are sent from the
corresponding senders below. Additionally, 'X' and 'O' symbols are used
to node where join builds arrive.

By default, fragments are printed with nodes grouped within each fragment
for easier timing analysis. A checkbox is provided that changes the
ordering so that nodes are printed in plan order similar to the Summary
tab.

The length of each colored phase interval depicts the longest instance
time while instances that complete earlier are shown as grey lines within
the phase interval.

On the server side, the query profile handler has been modified to return
the profile JSON as a document element instead of document content as it
is for JSON profile download. This allows the profile to be accessed in
full using scripts in templates.

Dense profiles will not render correctly as they currently do not contain
the necessary timestamps. This limitation will be addressed at a later
time with corresponding changes to the dense profile.

Example output is attached to the IMPALA-11970 ticket

Testing: Manual validation of rendering using TPC-H and TPC-DS queries
Will look into adding tests to test_web_pages.py and
test_observability.py in a subsequent patch

Change-Id: I8b5826107af0f5a7fe306cb986a875cff261d9db